[PRD-5458] Improve MS Excel export time by switching to streaming version of POI API

+++ engine/core/src/main/java/org/pentaho/reporting/engine/classic/core/modules/output/table/xls/helper/ExcelFontFactory.java
@@ -61,7 +61,7 @@ public class ExcelFontFactory {
     // read the fonts from the workbook ...
     // Funny one: Please note that the layout will be broken if the first
     // font is not 'Arial 10'.
-    final short numberOfFonts = this.workbook.getNumberOfFonts();
+    final int numberOfFonts = this.workbook.getNumberOfFonts();
     for ( int i = 0; i < numberOfFonts; i++ ) {
       final Font font = workbook.getFontAt( (short) i );
       this.fonts.put( new HSSFFontWrapper( font ), font );
